third_party/blink/web_tests/platform/generic/external/wpt/html/anonymous-iframe/local-storage-initial-empty-document.tentative.https.window-expected.txt

This is a testharness.js-based test.
FAIL Local storage is correctly partitioned with regards to anonymous iframe in initial empty documents. assert_equals: expected "" but got "value_E"
Harness: the test ran to completion.
